requirements.txt
flask==3.0.3
flask-cors==4.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
uvicorn==0.30.1

